Add Oracle update for release feeds
