+++ althea_kernel_interface/src/lib.rs
@@ -20,6 +20,7 @@ mod delete_tunnel;
 mod dns;
 mod exit_client_tunnel;
 mod exit_server_tunnel;
+pub mod file_io;
 mod fs_sync;
 mod get_neighbors;
 mod interface_tools;
@@ -31,6 +32,7 @@ mod link_local_tools;
 mod manipulate_uci;
 pub mod open_tunnel;
 mod openwrt_ubus;
+pub mod opkg_feeds;
 mod ping_check;
 mod set_system_password;
 mod setup_wg_if;
