+++ althea_types/src/lib.rs
@@ -1,5 +1,7 @@
 #[macro_use]
 extern crate serde_derive;
+#[macro_use]
+extern crate failure;
 
 extern crate arrayvec;
